refactor: extract types for Telegram API calls
